all modules created with controllers

+++ src/main/java/com/niit/collaborationbackend/dao/BlogDAOImpl.java
@@ -2,17 +2,17 @@ package com.niit.collaborationbackend.dao;
 
 import java.util.List;
 
-import org.hibernate.Query;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
+import org.springframework.transaction.annotation.Transactional;
+import org.hibernate.query.Query;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 import com.niit.collaborationbackend.model.Blog;
 
-@Repository("BlogDAO")
+//@Repository("blogDAOImpl")
 public class BlogDAOImpl implements BlogDAO {
 
 	@Autowired
@@ -24,20 +24,21 @@ public class BlogDAOImpl implements BlogDAO {
 	}
 
 	@Transactional
-	public void addBlog(Blog blog) {
+	public void addBlog(Blog b) {
 		Session session = this.sessionFactory.openSession();
 		Transaction tx = session.beginTransaction();
-		session.save(blog);
+		session.save(b);
 		tx.commit();
-		session.close();
+//		session.close();
 	}
 
 	@Transactional
 	public Blog getBlogByBlogId(int id) {
-		Session session = sessionFactory.openSession();
+		Session session = sessionFactory.getCurrentSession();
 		// System.out.print(id);
 		Blog b = (Blog) session.load(Blog.class, new Integer(id));
-		session.close();
+		System.out.println("get blog by id :" + b);
+		// session.close();
 		return b;
 	}
 
@@ -46,25 +47,73 @@ public class BlogDAOImpl implements BlogDAO {
 		Session session = sessionFactory.openSession();
 		Query query = session.createQuery("from Blog");
 		@SuppressWarnings("unchecked")
-		List<Blog> BlogsDetail = query.list();
+		List<Blog> blogsDetail = query.list();
 
-		return BlogsDetail;
+		return blogsDetail;
 	}
 
+	@SuppressWarnings("deprecation")
 	@Transactional
 	public Blog getBlogByBlogname(String Title) {
 		Session session = sessionFactory.openSession();
 
-		Query query = session.createQuery("from Blog where Title = ?");
+		Query query = session.createQuery("from Blog where blogname = ?");
 		query.setString(0, Title);
 
 		return (Blog) query.uniqueResult();
 	}
 
 	@Transactional
-	public void DeleteBlog(Blog blog) {
-		Session session = sessionFactory.openSession();
-		session.delete(blog);
-		session.close();
+	public void deleteBlog(Blog blog) {
+		Session session = this.sessionFactory.openSession();
+		Transaction tx = session.beginTransaction();
+		if (null != blog) {
+			session.delete(blog);
+		}
+		// logger.info("Product deleted successfully, product details="+p);
+		tx.commit();
+//		session.close();
 	}
+
+	@Transactional
+	public Blog updateBlog(int id , Blog blog) {
+		Session session = this.sessionFactory.openSession();
+		
+		Transaction tx = session.beginTransaction();
+		session.update(blog);
+		tx.commit();
+		Blog b = (Blog) session.load(Blog.class, new Integer(id));
+//		session.close();
+		// logger.info("Product updated successfully, Product Details="+p);
+		return b;
+	}
+	// public void storeFile(Blog blog) {
+	//
+	// MultipartFile file= blog.getFile();
+	//
+	// if (!file.isEmpty()) {
+	//
+	// try{
+	// byte[] bytes =file.getBytes();
+	// System.out.println(file.getOriginalFilename());
+	//
+	// String BASE_PATH="E:/";
+	// File serverFile = new
+	// File(BASE_PATH+p.getProduct_category()+"/"+blog.getImage());
+	// serverFile.createNewFile();
+	// BufferedOutputStream stream = new BufferedOutputStream(
+	// new FileOutputStream(serverFile));
+	// stream.write(bytes);
+	// stream.close();
+	// System.out.println("img stored");
+	// }
+	// catch(Exception ex)
+	// {
+	// System.out.println(ex);
+	// }
+	//
+	// }
+	// }
+
+	
 }
